correção exclusões

+++ doctrainingapp/views_pack/views_area.py
@@ -47,13 +47,18 @@ def area_edit(request, pk, template_name='area-edit.html'):
 @staff_member_required
 def area_delete(request, pk, template_name='area-delete.html'):
     area = get_object_or_404(Area, pk=pk)
-    try:
-        salas = Sala.objects.filter(area=area)
-        if salas:
-            messages.error(request, 'Erro! Area você não pode deletar essa Area pois existe salas contidas nela...')
-            # return redirect('/areas')
-            return redirect(reverse_lazy("doctrainingapp:areas_list"))
-    except:
+    salas = Sala.objects.filter(area=area)
+    fases = Fase.objects.filter(area=area)
+    if salas:
+        messages.error(request, 'Erro! Area você não pode deletar essa Area pois existe salas contidas nela...')
+        # return redirect('/areas')
+        return redirect(reverse_lazy("doctrainingapp:areas_list"))
+    if fases:
+        messages.error(request, 'Erro! Area você não pode deletar essa Area pois existe fases contidas nela...')
+        # return redirect('/areas')
+        return redirect(reverse_lazy("doctrainingapp:areas_list"))
+    elif request.method == 'POST':
+        print('xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx')
         area.delete()
         # return redirect('/areas')
         return redirect(reverse_lazy("doctrainingapp:areas_list"))
